size in floats for batik

+++ src/main/java/guru/nidi/graphviz/engine/Rasterizer.java
@@ -105,10 +105,10 @@ public enum Rasterizer {
                 };
                 final TranscodingHints hints = new TranscodingHints(t.getTranscodingHints());
                 if (graphviz.width != 0) {
-                    hints.put(ImageTranscoder.KEY_WIDTH, graphviz.scale * graphviz.width);
+                    hints.put(ImageTranscoder.KEY_WIDTH, (float) graphviz.scale * graphviz.width);
                 }
                 if (graphviz.height != 0) {
-                    hints.put(ImageTranscoder.KEY_HEIGHT, graphviz.scale * graphviz.height);
+                    hints.put(ImageTranscoder.KEY_HEIGHT, (float) graphviz.scale * graphviz.height);
                 }
                 t.setTranscodingHints(hints);
